Added transforms and uniform buffers
* Added functions for creating transform matrices
* Added uniform buffer class
* Fixed some warning about alignment (u32 -> u64) and redeclaration of
macros
* Added support for specifying winding order and cull mode when creation
pipeline states.

+++ olivine/src/olivine/render/api/buffer.cpp
@@ -45,7 +45,7 @@ Buffer::Buffer(const CreateInfo& createInfo)
 
 // -------------------------------------------------------------------------- //
 
-Buffer::Buffer(u64 size, Usage usages, HeapKind heapKind, u32 alignment)
+Buffer::Buffer(u64 size, Usage usages, HeapKind heapKind, u64 alignment)
   : mSize(size)
   , mAlignment(alignment)
 {
